requirements.txt
PyQt5
numpy
djitellopy
pillow